+++ tasks.py
@@ -1,57 +1,27 @@
+from __future__ import print_function
+
 import os
 
-import inquirer
+from compas_invocations2 import build, docs, mkdocs, style, tests
+from invoke.collection import Collection
 from invoke.tasks import task
-from rich.console import Console
-from rich.panel import Panel
 
-console = Console()
+ns = Collection(
+    docs.help,
+    style.check,
+    style.lint,
+    style.format,
+    mkdocs.docs,
+    tests.test,
+    tests.testdocs,
+    tests.testcodeblocks,
+    build.prepare_changelog,
+    build.clean,
+    build.release,
+)
 
-
-@task
-def rundev(c):
-    c.run("cd frontend/compas_threejs && npm run dev")
-
-
-@task
-def buildstatic(c):
-    with c.cd("frontend/compas_threejs"):
-        c.run("npm run build")
-
-
-@task
-def example(c):
-    """
-    Run an example from the examples folder.
-    """
-    examples_path = "examples"
-    examples = [
-        f for f in os.listdir(examples_path) if f.endswith(".py") and f != "__init__.py"
-    ]
-
-    if not examples:
-        console.print("[bold red]No examples found.[/bold red]")
-        return
-
-    questions = [
-        inquirer.List(
-            "example",
-            message="Which example would you like to run?",
-            choices=examples,
-        ),
-    ]
-
-    try:
-        console.print(
-            Panel.fit(
-                "[bold green]Select an Example to Run[/bold green]",
-                border_style="green",
-            )
-        )
-        answers = inquirer.prompt(questions)
-        if answers:
-            example_file = answers["example"]
-            console.print(f"\n[bold green]Running {example_file}...[/bold green]\n")
-            c.run(f"python examples/{example_file}")
-    except KeyboardInterrupt:
-        console.print("\n[bold yellow]Example selection cancelled.[/bold yellow]")
+ns.configure(
+    {
+        "base_folder": os.path.dirname(__file__),
+    }
+)